Moved CholeskySuiteSparse to a more appropriate namespace
